requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
drf-yasg==1.21.5
django-rest-swagger==2.2.0